rename ConvertToQuotes
