+++ docs/utilities.md
@@ -82,19 +82,19 @@ Currently, `.Condense()` is only available on a select few indicators.  If you f
 
 ### Convert to quotes
 
-`results.ConvertToQuotes()` will transform indicator results back into an `IEnumerable<Quote>` so it can be re-used to generate an [indicator of indicators]({{site.baseurl}}/guide/#generating-indicator-of-indicators).
+`results.ToQuotes()` will transform indicator results back into an `IEnumerable<Quote>` so it can be re-used to generate an [indicator of indicators]({{site.baseurl}}/guide/#generating-indicator-of-indicators).
 
 ```csharp
 // example: an RSI of Renko bricks
 IEnumerable<RsiResult> results
   = quotes.GetRenko(..)
-    .ConvertToQuotes()
+    .ToQuotes()
     .GetRsi(14);
 ```
 
-Currently, `.ConvertToQuotes` is only available on a select few indicators.  If you find an indicator that is a good candidate for this utility, please [submit an Issue]({{site.github.repository_url}}/issues).
+Currently, `.ToQuotes` is only available on a select few indicators.  If you find an indicator that is a good candidate for this utility, please [submit an Issue]({{site.github.repository_url}}/issues).
 
-:warning: WARNING! In many cases, `.ConvertToQuotes` will remove any `null` results -- this will produce fewer records than are in `quotes`.
+:warning: WARNING! In many cases, `.ToQuotes` will remove any `null` results -- this will produce fewer records than are in `quotes`.
 
 ### Find indicator result by date
