+++ src/m-r/Rsi/Rsi.Utilities.cs
@@ -6,7 +6,7 @@ public static partial class Indicator
     // convert to quotes
     /// <include file='../../_common/Results/info.xml' path='info/type[@name="Convert"]/*' />
     ///
-    public static IEnumerable<Quote> ConvertToQuotes(
+    public static IEnumerable<Quote> ToQuotes(
         this IEnumerable<RsiResult> results)
     {
         return results
