+++ tests/indicators/e-k/HeikinAshi/HeikinAshi.Tests.cs
@@ -26,25 +26,6 @@ public class HeikinAshi : TestBase
     }
 
     [TestMethod]
-    public void ConvertToQuotes()
-    {
-        List<Quote> newQuotes = quotes.GetHeikinAshi()
-            .ConvertToQuotes()
-            .ToList();
-
-        // assertions
-
-        Assert.AreEqual(502, newQuotes.Count);
-
-        Quote q = newQuotes[501];
-        Assert.AreEqual(241.3018m, Math.Round(q.Open, 4));
-        Assert.AreEqual(245.54m, Math.Round(q.High, 4));
-        Assert.AreEqual(241.3018m, Math.Round(q.Low, 4));
-        Assert.AreEqual(244.6525m, Math.Round(q.Close, 4));
-        Assert.AreEqual(147031456m, q.Volume);
-    }
-
-    [TestMethod]
     public void UseAsQuotes()
     {
         IEnumerable<HeikinAshiResult> haQuotes = quotes.GetHeikinAshi();
